refactor(PostBean): 変数をint型に変更

+++ app/src/main/java/com/example/snap_develop/bean/PostBean.java
@@ -17,7 +17,7 @@ public class PostBean implements Serializable {
     private String message;
     private String photoName;
     private Bitmap photo;
-    private Long goodCount;
+    private int goodCount;
     private LatLng latLng;
     private Date datetime;
     private String strDatetime;
